+++ CHANGELOG.md
@@ -12,6 +12,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - #957 - Adds telemetry for the context menu
 - #1036 - Adds telemetry for Find in Page
 - #1049 - Add style for progress bar with gradient drawable
-
+- #1165 - Added doorhanger to the toolbar
 ### Changed
 ### Removed